requirements.txt
stomp.py==8.1.0
paho-mqtt==1.6.1
